docs/source/Developer-Tools/Cloud-Development-Kit-Root/03-Deploy-CDK-Stack-To-Multiple-Account/requirements.txt
aws-cdk-lib==2.130.0
constructs==10.2.70
